+++ src/camera/arc_ball.rs
@@ -16,42 +16,42 @@ use event;
 #[deriving(Clone, ToStr)]
 pub struct ArcBall {
     /// The focus point.
-    priv at:    Vec3<f64>,
+    priv at:    Vec3<f32>,
     /// Yaw of the camera (rotation along the y axis).
-    priv yaw:   f64,
+    priv yaw:   f32,
     /// Pitch of the camera (rotation along the x axis).
-    priv pitch: f64,
+    priv pitch: f32,
     /// Distance from the camera to the `at` focus point.
-    priv dist:  f64,
+    priv dist:  f32,
 
     /// Increment of the yaw per unit mouse movement. The default value is 0.005.
-    priv yaw_step:   f64,
+    priv yaw_step:   f32,
     /// Increment of the pitch per unit mouse movement. The default value is 0.005.
-    priv pitch_step: f64,
+    priv pitch_step: f32,
     /// Increment of the distance per unit scrolling. The default value is 40.0.
-    priv dist_step:  f64,
+    priv dist_step:  f32,
 
-    priv fov:        f64,
-    priv znear:      f64,
-    priv zfar:       f64,
-    priv projection:      Mat4<f64>,
-    priv proj_view:       Mat4<f64>,
-    priv inv_proj_view:   Mat4<f64>,
-    priv last_cursor_pos: Vec2<f64>
+    priv fov:        f32,
+    priv znear:      f32,
+    priv zfar:       f32,
+    priv projection:      Mat4<f32>,
+    priv proj_view:       Mat4<f32>,
+    priv inv_proj_view:   Mat4<f32>,
+    priv last_cursor_pos: Vec2<f32>
 }
 
 impl ArcBall {
     /// Create a new arc-ball camera.
-    pub fn new(eye: Vec3<f64>, at: Vec3<f64>) -> ArcBall {
-        ArcBall::new_with_frustrum(45.0f64.to_radians(), 0.1, 1024.0, eye, at)
+    pub fn new(eye: Vec3<f32>, at: Vec3<f32>) -> ArcBall {
+        ArcBall::new_with_frustrum(45.0f32.to_radians(), 0.1, 1024.0, eye, at)
     }
 
     /// Creates a new arc ball camera with default sensitivity values.
-    pub fn new_with_frustrum(fov:    f64,
-                             znear:  f64,
-                             zfar:   f64,
-                             eye:    Vec3<f64>,
-                             at:     Vec3<f64>) -> ArcBall {
+    pub fn new_with_frustrum(fov:    f32,
+                             znear:  f32,
+                             zfar:   f32,
+                             eye:    Vec3<f32>,
+                             at:     Vec3<f32>) -> ArcBall {
         let mut res = ArcBall {
             at:         Vec3::new(0.0, 0.0, 0.0),
             yaw:        0.0,
@@ -75,17 +75,17 @@ impl ArcBall {
     }
 
     /// The point the arc-ball is looking at.
-    pub fn at(&self) -> Vec3<f64> {
+    pub fn at(&self) -> Vec3<f32> {
         self.at
     }
 
     /// The arc-ball camera `yaw`.
-    pub fn yaw(&self) -> f64 {
+    pub fn yaw(&self) -> f32 {
         self.yaw
     }
 
     /// Sets the camera `yaw`. Change this to modify the rotation along the `up` axis.
-    pub fn set_yaw(&mut self, yaw: f64) {
+    pub fn set_yaw(&mut self, yaw: f32) {
         self.yaw = yaw;
 
         self.update_restrictions();
@@ -93,12 +93,12 @@ impl ArcBall {
     }
 
     /// The arc-ball camera `pitch`.
-    pub fn pitch(&self) -> f64 {
+    pub fn pitch(&self) -> f32 {
         self.pitch
     }
 
     /// Sets the camera `pitch`.
-    pub fn set_pitch(&mut self, pitch: f64) {
+    pub fn set_pitch(&mut self, pitch: f32) {
         self.pitch = pitch;
 
         self.update_restrictions();
@@ -106,12 +106,12 @@ impl ArcBall {
     }
 
     /// The distance from the camera position to its view point.
-    pub fn dist(&self) -> f64 {
+    pub fn dist(&self) -> f32 {
         self.dist
     }
 
     /// Move the camera such that it is at a given distance from the view point.
-    pub fn set_dist(&mut self, dist: f64) {
+    pub fn set_dist(&mut self, dist: f32) {
         self.dist = dist;
 
         self.update_restrictions();
@@ -119,7 +119,7 @@ impl ArcBall {
     }
 
     /// Move and orient the camera such that it looks at a specific point.
-    pub fn look_at_z(&mut self, eye: Vec3<f64>, at: Vec3<f64>) {
+    pub fn look_at_z(&mut self, eye: Vec3<f32>, at: Vec3<f32>) {
         let dist  = na::norm(&(eye - at));
         let pitch = ((eye.y - at.y) / dist).acos();
         let yaw   = atan2(eye.z - at.z, eye.x - at.x);
@@ -141,13 +141,13 @@ impl ArcBall {
             self.pitch = 0.0001
         }
 
-        let _pi: f64 = Real::pi();
+        let _pi: f32 = Real::pi();
         if (self.pitch > _pi - 0.0001) {
             self.pitch = _pi - 0.0001
         }
     }
 
-    fn handle_left_button_displacement(&mut self, dpos: &Vec2<f64>) {
+    fn handle_left_button_displacement(&mut self, dpos: &Vec2<f32>) {
         self.yaw   = self.yaw   + dpos.x * self.yaw_step;
         self.pitch = self.pitch - dpos.y * self.pitch_step;
 
@@ -155,7 +155,7 @@ impl ArcBall {
         self.update_projviews();
     }
 
-    fn handle_right_button_displacement(&mut self, dpos: &Vec2<f64>) {
+    fn handle_right_button_displacement(&mut self, dpos: &Vec2<f32>) {
         let eye       = self.eye();
         let dir       = na::normalize(&(self.at - eye));
         let tangent   = na::normalize(&na::cross(&Vec3::y(), &dir));
@@ -166,8 +166,8 @@ impl ArcBall {
         self.update_projviews();
     }
 
-    fn handle_scroll(&mut self, off: f64) {
-        self.dist = self.dist + self.dist_step * (off as f64) / 120.0;
+    fn handle_scroll(&mut self, off: f32) {
+        self.dist = self.dist + self.dist_step * (off) / 120.0;
         self.update_restrictions();
         self.update_projviews();
     }
@@ -179,18 +179,18 @@ impl ArcBall {
 }
 
 impl Camera for ArcBall {
-    fn clip_planes(&self) -> (f64, f64) {
+    fn clip_planes(&self) -> (f32, f32) {
         (self.znear, self.zfar)
     }
 
-    fn view_transform(&self) -> Iso3<f64> {
-        let mut id: Iso3<f64> = na::one();
+    fn view_transform(&self) -> Iso3<f32> {
+        let mut id: Iso3<f32> = na::one();
         id.look_at_z(&self.eye(), &self.at, &Vec3::y());
 
         id
     }
 
-    fn eye(&self) -> Vec3<f64> {
+    fn eye(&self) -> Vec3<f32> {
         let px = self.at.x + self.dist * self.yaw.cos() * self.pitch.sin();
         let py = self.at.y + self.dist * self.pitch.cos();
         let pz = self.at.z + self.dist * self.yaw.sin() * self.pitch.sin();
@@ -201,7 +201,7 @@ impl Camera for ArcBall {
     fn handle_event(&mut self, window: &glfw::Window, event: &event::Event) {
         match *event {
             event::CursorPos(x, y) => {
-                let curr_pos = Vec2::new(x as f64, y as f64);
+                let curr_pos = Vec2::new(x, y);
 
                 if window.get_mouse_button(glfw::MouseButtonLeft) == glfw::Press {
                     let dpos = curr_pos - self.last_cursor_pos;
@@ -228,11 +228,11 @@ impl Camera for ArcBall {
         }
     }
 
-    fn transformation(&self) -> Mat4<f64> {
+    fn transformation(&self) -> Mat4<f32> {
         self.proj_view
     }
 
-    fn inv_transformation(&self) -> Mat4<f64> {
+    fn inv_transformation(&self) -> Mat4<f32> {
         self.inv_proj_view
     }
