+++ src/camera/first_person_stereo.rs
@@ -18,50 +18,50 @@ mod error;
 #[deriving(ToStr)]
 pub struct FirstPersonStereo {
     /// The camera position
-    priv eye:        Vec3<f64>,
-    priv eye_left:   Vec3<f64>,
-    priv eye_right:  Vec3<f64>,
+    priv eye:        Vec3<f32>,
+    priv eye_left:   Vec3<f32>,
+    priv eye_right:  Vec3<f32>,
 
     /// Inter Pupilary Distance
-    priv ipd:        f64,
+    priv ipd:        f32,
 
     /// Yaw of the camera (rotation along the y axis).
-    priv yaw:        f64,
+    priv yaw:        f32,
     /// Pitch of the camera (rotation along the x axis).
-    priv pitch:      f64,
+    priv pitch:      f32,
 
     /// Increment of the yaw per unit mouse movement. The default value is 0.005.
-    priv yaw_step:   f64,
+    priv yaw_step:   f32,
     /// Increment of the pitch per unit mouse movement. The default value is 0.005.
-    priv pitch_step: f64,
+    priv pitch_step: f32,
     /// Increment of the translation per arrow press. The default value is 0.1.
-    priv move_step:  f64,
+    priv move_step:  f32,
 
     /// Low level datas
-    priv fov:        f64,
-    priv znear:      f64,
-    priv zfar:       f64,
-    priv projection:      Mat4<f64>,
-    priv proj_view:       Mat4<f64>,
-    priv proj_view_left:  Mat4<f64>,
-    priv proj_view_right: Mat4<f64>,
-    priv inv_proj_view:   Mat4<f64>,
-    priv last_cursor_pos: Vec2<f64>
+    priv fov:        f32,
+    priv znear:      f32,
+    priv zfar:       f32,
+    priv projection:      Mat4<f32>,
+    priv proj_view:       Mat4<f32>,
+    priv proj_view_left:  Mat4<f32>,
+    priv proj_view_right: Mat4<f32>,
+    priv inv_proj_view:   Mat4<f32>,
+    priv last_cursor_pos: Vec2<f32>
 }
 
 impl FirstPersonStereo {
     /// Creates a first person camera with default sensitivity values.
-    pub fn new(eye: Vec3<f64>, at: Vec3<f64>, ipd: f64) -> FirstPersonStereo {
-        FirstPersonStereo::new_with_frustrum(45.0f64.to_radians(), 0.1, 1024.0, eye, at, ipd)
+    pub fn new(eye: Vec3<f32>, at: Vec3<f32>, ipd: f32) -> FirstPersonStereo {
+        FirstPersonStereo::new_with_frustrum(45.0f32.to_radians(), 0.1, 1024.0, eye, at, ipd)
     }
 
     /// Creates a new first person camera with default sensitivity values.
-    pub fn new_with_frustrum(fov:    f64,
-                             znear:  f64,
-                             zfar:   f64,
-                             eye:    Vec3<f64>,
-                             at:     Vec3<f64>,
-                             ipd:    f64) -> FirstPersonStereo {
+    pub fn new_with_frustrum(fov:    f32,
+                             znear:  f32,
+                             zfar:   f32,
+                             eye:    Vec3<f32>,
+                             at:     Vec3<f32>,
+                             ipd:    f32) -> FirstPersonStereo {
         let mut res = FirstPersonStereo {
             eye:           Vec3::new(0.0, 0.0, 0.0),
             // left & right are initially wrong, don't take ipd into accound
@@ -91,7 +91,7 @@ impl FirstPersonStereo {
 
 
     /// Changes the orientation and position of the camera to look at the specified point.
-    pub fn look_at_z(&mut self, eye: Vec3<f64>, at: Vec3<f64>) {
+    pub fn look_at_z(&mut self, eye: Vec3<f32>, at: Vec3<f32>) {
         let dist  = na::norm(&(eye - at));
 
         let pitch = ((at.y - eye.y) / dist).acos();
@@ -104,7 +104,7 @@ impl FirstPersonStereo {
     }
 
     /// The point the camera is looking at.
-    pub fn at(&self) -> Vec3<f64> {
+    pub fn at(&self) -> Vec3<f32> {
         let ax = self.eye.x + self.yaw.cos() * self.pitch.sin();
         let ay = self.eye.y + self.pitch.cos();
         let az = self.eye.z + self.yaw.sin() * self.pitch.sin();
@@ -117,14 +117,14 @@ impl FirstPersonStereo {
             self.pitch = 0.0001
         }
 
-        let _pi: f64 = Real::pi();
+        let _pi: f32 = Real::pi();
         if (self.pitch > _pi - 0.0001) {
             self.pitch = _pi - 0.0001
         }
     }
 
     #[doc(hidden)]
-    pub fn handle_left_button_displacement(&mut self, dpos: &Vec2<f64>) {
+    pub fn handle_left_button_displacement(&mut self, dpos: &Vec2<f32>) {
         self.yaw   = self.yaw   + dpos.x * self.yaw_step;
         self.pitch = self.pitch + dpos.y * self.pitch_step;
 
@@ -137,21 +137,21 @@ impl FirstPersonStereo {
         // up is always y
         let dir       = na::normalize(&(self.at() - self.eye));
         let tangent   = na::normalize(&na::cross(&Vec3::y(), &dir));
-        self.eye_left = self.eye - tangent * (self.ipd / 2.0 as f64);
-        self.eye_right = self.eye + tangent * (self.ipd / 2.0 as f64);
+        self.eye_left = self.eye - tangent * (self.ipd / 2.0);
+        self.eye_right = self.eye + tangent * (self.ipd / 2.0);
         //println(fmt!("eye_left = %f,%f,%f", self.eye_left.x as float, self.eye_left.y as float, self.eye_left.z as float));
         //println(fmt!("eye_right = %f,%f,%f", self.eye_right.x as float, self.eye_right.y as float, self.eye_right.z as float));
         // TODO: verify with an assert or something that the distance between the eyes is ipd, just to make me feel good.
     }
 
     #[doc(hidden)]
-    pub fn handle_right_button_displacement(&mut self, dpos: &Vec2<f64>) {
+    pub fn handle_right_button_displacement(&mut self, dpos: &Vec2<f32>) {
         let at        = self.at();
         let dir       = na::normalize(&(at - self.eye));
         let tangent   = na::normalize(&na::cross(&Vec3::y(), &dir));
         let bitangent = na::cross(&dir, &tangent);
 
-        self.eye = self.eye + tangent * (0.01 * dpos.x as f64 / 10.0) + bitangent * (0.01 * dpos.y as f64 / 10.0);
+        self.eye = self.eye + tangent * (0.01 * dpos.x / 10.0) + bitangent * (0.01 * dpos.y / 10.0);
         // TODO: ugly - should move eye update to where eye_left & eye_right are updated
         self.update_eyes_location();
         self.update_restrictions();
@@ -159,8 +159,8 @@ impl FirstPersonStereo {
     }
 
     #[doc(hidden)]
-    pub fn handle_scroll(&mut self, yoff: f64) {
-        let front: Vec3<f64> = self.view_transform().rotate(&Vec3::z());
+    pub fn handle_scroll(&mut self, yoff: f32) {
+        let front: Vec3<f32> = self.view_transform().rotate(&Vec3::z());
 
         self.eye = self.eye + front * (self.move_step * yoff);
 
@@ -176,7 +176,7 @@ impl FirstPersonStereo {
         self.proj_view_right = self.projection * na::to_homogeneous(&na::inv(&self.view_transform_right()).unwrap());
     }
 
-    fn transformation_eye(&self, eye: uint) -> Mat4<f64> {
+    fn transformation_eye(&self, eye: uint) -> Mat4<f32> {
         match eye {
             0u => self.proj_view_left,
             1u => self.proj_view_right,
@@ -185,28 +185,28 @@ impl FirstPersonStereo {
     }
 
     /// The left eye camera view transformation
-    fn view_transform_left(&self) -> Iso3<f64> {
-        let mut id: Iso3<f64> = One::one();
+    fn view_transform_left(&self) -> Iso3<f32> {
+        let mut id: Iso3<f32> = One::one();
         id.look_at_z(&self.eye_left, &self.at(), &Vec3::y());
 
         id
     }
 
     /// The right eye camera view transformation
-    fn view_transform_right(&self) -> Iso3<f64> {
-        let mut id: Iso3<f64> = One::one();
+    fn view_transform_right(&self) -> Iso3<f32> {
+        let mut id: Iso3<f32> = One::one();
         id.look_at_z(&self.eye_right, &self.at(), &Vec3::y());
 
         id
     }
 
     /// return Inter Pupilary Distance
-    pub fn ipd(&self) -> f64 {
+    pub fn ipd(&self) -> f32 {
         self.ipd
     }
     
     /// change Inter Pupilary Distance
-    pub fn set_ipd(&mut self, ipd: f64) {
+    pub fn set_ipd(&mut self, ipd: f32) {
         self.ipd = ipd;
 
         self.update_eyes_location();
@@ -217,13 +217,13 @@ impl FirstPersonStereo {
 }
 
 impl Camera for FirstPersonStereo {
-    fn clip_planes(&self) -> (f64, f64) {
+    fn clip_planes(&self) -> (f32, f32) {
         (self.znear, self.zfar)
     }
 
     /// The imaginary middle eye camera view transformation (i-e transformation without projection).
-    fn view_transform(&self) -> Iso3<f64> {
-        let mut id: Iso3<f64> = One::one();
+    fn view_transform(&self) -> Iso3<f32> {
+        let mut id: Iso3<f32> = One::one();
         id.look_at_z(&self.eye, &self.at(), &Vec3::y());
 
         id
@@ -232,7 +232,7 @@ impl Camera for FirstPersonStereo {
     fn handle_event(&mut self, window: &glfw::Window, event: &event::Event) {
         match *event {
             event::CursorPos(x, y) => {
-                let curr_pos = Vec2::new(x as f64, y as f64);
+                let curr_pos = Vec2::new(x, y);
 
                 if window.get_mouse_button(glfw::MouseButtonLeft) == glfw::Press {
                     let dpos = curr_pos - self.last_cursor_pos;
@@ -255,22 +255,22 @@ impl Camera for FirstPersonStereo {
         }
     }
 
-    fn eye(&self) -> Vec3<f64> {
+    fn eye(&self) -> Vec3<f32> {
         self.eye
     }
 
-    fn transformation(&self) -> Mat4<f64> {
+    fn transformation(&self) -> Mat4<f32> {
         self.proj_view
     }
 
-    fn inv_transformation(&self) -> Mat4<f64> {
+    fn inv_transformation(&self) -> Mat4<f32> {
         self.inv_proj_view
     }
 
     fn update(&mut self, window: &glfw::Window) {
         let t                = self.view_transform();
-        let front: Vec3<f64> = t.rotate(&Vec3::z());
-        let right: Vec3<f64> = t.rotate(&Vec3::x());
+        let front: Vec3<f32> = t.rotate(&Vec3::z());
+        let right: Vec3<f32> = t.rotate(&Vec3::x());
 
         if window.get_key(glfw::KeyUp) == glfw::Press {
             self.eye = self.eye + front * self.move_step
