+++ src/event.rs
@@ -10,13 +10,13 @@ pub enum Event {
     /// Event triggered when a keyboard key is released.
     KeyReleased(glfw::Key),
     /// Event triggered when the window framebuffer is resized.
-    FramebufferSize(f64, f64),
+    FramebufferSize(f32, f32),
     /// Event triggered when a mouse button is pressed.
     ButtonPressed(glfw::MouseButton,  glfw::Modifiers),
     /// Event triggered when a mouse button is released.
     ButtonReleased(glfw::MouseButton, glfw::Modifiers),
     /// Event triggered when the cursor position changes.
-    CursorPos(f64, f64),
+    CursorPos(f32, f32),
     /// Event triggered when the mouse scrolls.
-    Scroll(f64, f64)
+    Scroll(f32, f32)
 }
